Improved code quality
tendrl-bug-id: Tendrl/monitoring-integration#325
