+++ tendrl/monitoring_integration/grafana/alert_utils.py
@@ -1,7 +1,7 @@
 import json
 
 from tendrl.monitoring_integration.grafana import constants
-from tendrl.monitoring_integration.grafana import dashboard
+from tendrl.monitoring_integration.grafana import dashboard_utils
 from tendrl.monitoring_integration.grafana import grafana_org_utils
 
 
@@ -13,9 +13,9 @@ def get_alert_dashboard(dashboard_name):
     dashboard_json = {}
     if grafana_org_utils.get_current_org_name()["name"] == \
             constants.ALERT_ORG:
-        dashboard_json = dashboard.get_dashboard(slug)
+        dashboard_json = dashboard_utils.get_dashboard(slug)
     elif switch_context(constants.ALERT_ORG):
-        dashboard_json = dashboard.get_dashboard(slug)
+        dashboard_json = dashboard_utils.get_dashboard(slug)
         # return to main org
         switch_context(constants.MAIN_ORG)
     return dashboard_json
@@ -36,21 +36,33 @@ def delete_alert_dashboard(dashboard_name):
     dashboard_json = {}
     if grafana_org_utils.get_current_org_name()["name"] == \
             constants.ALERT_ORG:
-        dashboard_json = dashboard.delete_dashboard(slug)
+        dashboard_json = dashboard_utils.delete_dashboard(slug)
     elif switch_context(constants.ALERT_ORG):
-        dashboard_json = dashboard.delete_dashboard(slug)
+        dashboard_json = dashboard_utils.delete_dashboard(slug)
         # return to main org
         switch_context(constants.MAIN_ORG)
     return dashboard_json
 
 
 def post_dashboard(alert_dashboard):
-        resp = None
-        if grafana_org_utils.get_current_org_name()["name"] == \
-                constants.ALERT_ORG:
-            resp = dashboard._post_dashboard(alert_dashboard)
-        elif switch_context(constants.ALERT_ORG):
-            resp = dashboard._post_dashboard(alert_dashboard)
-            # return to main org
-            switch_context(constants.MAIN_ORG)
-        return resp
+    resp = None
+    if grafana_org_utils.get_current_org_name()["name"] == \
+            constants.ALERT_ORG:
+        resp = dashboard_utils._post_dashboard(alert_dashboard)
+    elif switch_context(constants.ALERT_ORG):
+        resp = dashboard_utils._post_dashboard(alert_dashboard)
+        # return to main org
+        switch_context(constants.MAIN_ORG)
+    return resp
+
+
+def get_alert(alert_id):
+    resp = None
+    if grafana_org_utils.get_current_org_name()["name"] == \
+            constants.ALERT_ORG:
+        resp = dashboard_utils.get_alert(alert_id)
+    elif switch_context(constants.ALERT_ORG):
+        resp = dashboard_utils.get_alert(alert_id)
+        # return to main org
+        switch_context(constants.MAIN_ORG)
+    return resp.json()
