+++ tendrl/monitoring_integration/grafana/dashboard.py
@@ -1,117 +1,94 @@
 import json
-import maps
-import os
 import traceback
 
-from requests import delete
-from requests import get
-from requests import post
-from requests import put
-
+from tendrl.commons.utils import log_utils as logger
 from tendrl.monitoring_integration.grafana import constants
+from tendrl.monitoring_integration.grafana import dashboard_utils
+from tendrl.monitoring_integration.grafana import datasource
 from tendrl.monitoring_integration.grafana import exceptions
+from tendrl.monitoring_integration.grafana import grafana_org_utils
 from tendrl.monitoring_integration.grafana import utils
 
 
-def _post_dashboard(dashboard_json, authorization_key=None):
-    config = maps.NamedDict(NS.config.data)
-    if utils.port_open(config.grafana_port, config.grafana_host):
-        upload_str = json.dumps(dashboard_json)
-        if authorization_key:
-            new_header = constants.HEADERS
-            new_header["Authorization"] = "Bearer " + str(authorization_key)
-            response = post("http://{}:{}/api/dashboards/"
-                            "db".format(config.grafana_host,
-                                        config.grafana_port),
-                            headers=new_header,
-                            data=upload_str)
+def upload_default_dashboards():
+    dashboards = []
+    NS.config.data["credentials"] = utils.get_credentials()
+    try:
+        main_org_id = grafana_org_utils.get_org_id(constants.MAIN_ORG)
+        if main_org_id:
+            response = grafana_org_utils.switch_context(
+                json.loads(main_org_id)["id"]
+            )
+    except (exceptions.ConnectionFailedException, KeyError) as ex:
+        msg = (json.loads(main_org_id)).get(
+            "message", "Cannot connect to grafana")
+        logger.log("error", NS.get("publisher_id", None),
+                   {'message': msg})
+        raise ex
+    title = []
+    # create datasource
+    datasource.create()
+    dashboards = dashboard_utils.get_all_dashboards()
+    for dashboard_json in dashboards:
+        title.append(dashboard_json["uri"].split('/')[1])
+
+    for dashboard_json in NS.config.data["dashboards"]:
+        if dashboard_json in title:
+            msg = '\n' + "Dashboard " + str(dashboard_json) + \
+                  " already exists" + '\n'
+            logger.log("info", NS.get("publisher_id", None),
+                       {'message': msg})
+            continue
+        response = dashboard_utils.create_dashboard(dashboard_json)
+
+        if response.status_code == 200:
+            msg = '\n' + "Dashboard " + str(dashboard_json) + \
+                  " uploaded successfully" + '\n'
+            logger.log("info", NS.get("publisher_id", None),
+                       {'message': msg})
         else:
-            response = post("http://{}:{}/api/dashboards/"
-                            "db".format(config.grafana_host,
-                                        config.grafana_port),
-                            headers=constants.HEADERS,
-                            auth=config.credentials,
-                            data=upload_str)
-        return response
-    else:
+            msg = ("Dashboard {0} upload failed. Error code: {1} ," +
+                   "Error message: " + "{2} ").format(
+                       str(dashboard_json),
+                       str(response.status_code),
+                       str(get_message_from_response(response)))
+            logger.log("error", NS.get("publisher_id", None),
+                       {'message': msg})
+    try:
+        dashboard_json = dashboard_utils.get_dashboard(
+            NS.config.data["home_dashboard"])
+
+        if 'dashboard' in dashboard_json:
+            dashboard_id = dashboard_json.get('dashboard').get('id')
+            response = dashboard_utils.set_home_dashboard(dashboard_id)
+
+            response = dashboard_utils.set_home_dashboard(dashboard_id)
+            if response.status_code == 200:
+                msg = '\n' + "Dashboard " + \
+                      str(NS.config.data["home_dashboard"]) + \
+                      " is set as home dashboard" + '\n'
+                logger.log("info", NS.get("publisher_id", None),
+                           {'message': msg})
+        else:
+            msg = '\n' + str(dashboard_json.get('message')) + '\n'
+            logger.log("info", NS.get("publisher_id", None),
+                       {'message': msg})
+    except exceptions.ConnectionFailedException as ex:
+        traceback.print_exc()
+        logger.log("error", NS.get("publisher_id", None),
+                   {'message': str(ex)})
         raise exceptions.ConnectionFailedException
 
 
-def get_dashboard(dashboard_name):
-    config = maps.NamedDict(NS.config.data)
-    if utils.port_open(config.grafana_port, config.grafana_host):
-        resp = get("http://{}:{}/api/dashboards/"
-                   "db/{}".format(config.grafana_host,
-                                  config.grafana_port,
-                                  dashboard_name),
-                   auth=config.credentials)
-    else:
-        raise exceptions.ConnectionFailedException
-    return resp.json()
+def get_message_from_response(response_data):
 
+    message = ""
+    try:
+        if isinstance(json.loads(response_data.content), list):
+            message = str(json.loads(response_data.content)[0]["message"])
+        else:
+            message = str(json.loads(response_data.content)["message"])
+    except (AttributeError, KeyError):
+        pass
 
-def delete_dashboard(dashboard_name):
-    config = maps.NamedDict(NS.config.data)
-    if utils.port_open(config.grafana_port, config.grafana_host):
-        resp = delete(
-            "http://{}:{}/api/dashboards/db/{}".format(
-                config.grafana_host,
-                config.grafana_port,
-                dashboard_name
-            ),
-            auth=config.credentials
-        )
-    else:
-        raise exceptions.ConnectionFailedException
-    return resp.json()
-
-
-def get_all_dashboards():
-    config = maps.NamedDict(NS.config.data)
-    if utils.port_open(config.grafana_port, config.grafana_host):
-        resp = get("http://{}:{}/api/search/"
-                   .format(config.grafana_host,
-                           config.grafana_port),
-                   auth=config.credentials)
-    else:
-        raise exceptions.ConnectionFailedException
-    return resp.json()
-
-
-def set_home_dashboard(dash_id):
-    config = maps.NamedDict(NS.config.data)
-    if utils.port_open(config.grafana_port, config.grafana_host):
-        resp = put('http://{}:{}/api/org/'
-                   'preferences'.format(config.grafana_host,
-                                        config.grafana_port),
-                   headers=constants.HEADERS,
-                   auth=config.credentials,
-                   data=json.dumps({"name": constants.MAIN_ORG,
-                                    "theme": "light",
-                                    "homeDashboardId": dash_id}))
-    else:
-        raise exceptions.ConnectionFailedException
-    return resp
-
-
-def create_dashboard(dashboard_name, dashboard_dir=None):
-    if not dashboard_dir:
-        dashboard_dir = "/etc/tendrl/monitoring-integration/grafana/dashboards"
-    dashboard_path = os.path.join(dashboard_dir,
-                                  "{}.json".format(dashboard_name))
-
-    if os.path.exists(dashboard_path):
-
-        dashboard_data = utils.fread(dashboard_path)
-
-        try:
-            dashboard_json = json.loads(dashboard_data)
-            response = _post_dashboard(dashboard_json)
-            return response
-
-        except exceptions.ConnectionFailedException:
-            traceback.print_stack()
-            raise exceptions.ConnectionFailedException
-
-    else:
-        raise exceptions.FileNotFoundException
+    return message
