+++ tendrl/monitoring_integration/tests/grafana/test_alert_organization.py
@@ -3,20 +3,20 @@ from etcd import Client
 import mock
 from mock import patch
 
-from tendrl.monitoring_integration.grafana import create_alert_organization
+from tendrl.monitoring_integration.grafana import alert_organization
 from tendrl.monitoring_integration.grafana import \
-    create_datasource
-from tendrl.monitoring_integration.grafana import \
-    create_notification_channel
+    datasource
 from tendrl.monitoring_integration.grafana import grafana_org_utils
+from tendrl.monitoring_integration.grafana import \
+    notification_channel
 from tendrl.monitoring_integration.tests import test_init
 
 
 @patch.object(etcd, "Client")
 @patch.object(Client, "write")
 @patch.object(grafana_org_utils, "switch_context")
-@patch.object(create_datasource, "create")
-@patch.object(create_notification_channel, "create_notification_channel")
+@patch.object(datasource, "create")
+@patch.object(notification_channel, "create_notification_channel")
 def test_create(create_notification, create_ds, context, write, client):
     write.return_value = etcd.Client()
     client.return_value = etcd.Client()
@@ -39,7 +39,7 @@ def test_create(create_notification, create_ds, context, write, client):
                     obj = NS.monitoring.objects.AlertOrganization(
                         auth_key="grafana_key")
                     load.return_value = obj
-                    create_alert_organization.create()
+                    alert_organization.create()
                     if not NS.config.data["org_id"] == 2:
                         raise AssertionError()
                     if not NS.config.data["grafana_auth_key"] == \
@@ -55,7 +55,7 @@ def test_create(create_notification, create_ds, context, write, client):
                     with patch.object(grafana_org_utils,
                                       "create_api_token") as api_token:
                         api_token.return_value = "grafana_key_new"
-                        create_alert_organization.create()
+                        alert_organization.create()
                         if not NS.config.data["org_id"] == 3:
                             raise AssertionError()
                         if not NS.config.data["grafana_auth_key"] ==  \
@@ -65,7 +65,7 @@ def test_create(create_notification, create_ds, context, write, client):
             org_utils.return_value = '{}'
             with patch.object(grafana_org_utils, "get_auth_keys") as auth_key:
                 auth_key.return_value = {}
-                create_alert_organization.create()
+                alert_organization.create()
                 if not NS.config.data["org_id"] is None:
                     raise AssertionError()
                 if not NS.config.data["grafana_auth_key"] is None:
