+++ tendrl/monitoring_integration/tests/grafana/test_datasource.py
@@ -1,8 +1,8 @@
 import maps
 import mock
 
-from tendrl.monitoring_integration.grafana import create_datasource
 from tendrl.monitoring_integration.grafana import datasource
+from tendrl.monitoring_integration.grafana import datasource_utils
 from tendrl.monitoring_integration.tests import test_init
 
 
@@ -43,42 +43,42 @@ def test_create():
     with mock.patch("tendrl.commons.utils.log_utils.log") as mock_log:
         test_init.init()
         with mock.patch.object(
-            datasource, "create_datasource", pass_create_datasource
+            datasource_utils, "create_datasource", pass_create_datasource
         ):
-            create_datasource.create()
+            datasource.create()
             mock_log.assert_called_with(
                 'info',
                 'monitoring_integration',
                 {'message': 'Datasource created successfully'}
             )
         with mock.patch.object(
-            datasource, "create_datasource", update_create_datasource
+            datasource_utils, "create_datasource", update_create_datasource
         ):
             with mock.patch.object(
-                datasource, "get_data_source", get_data_source_fail
+                datasource_utils, "get_data_source", get_data_source_fail
             ):
-                create_datasource.create()
+                datasource.create()
                 mock_log.assert_called_with(
                     'error',
                     'monitoring_integration',
                     {'message': 'Unable to find datasource id'}
                 )
             with mock.patch.object(
-                datasource, "get_data_source", get_data_source
+                datasource_utils, "get_data_source", get_data_source
             ):
                 with mock.patch.object(
-                    datasource, "update_datasource", update
+                    datasource_utils, "update_datasource", update
                 ):
-                    create_datasource.create()
+                    datasource.create()
                     mock_log.assert_called_with(
                         'info',
                         'monitoring_integration',
                         {'message': 'Datasource is updated successfully'}
                     )
                 with mock.patch.object(
-                    datasource, "update_datasource", fail_update
+                    datasource_utils, "update_datasource", fail_update
                 ):
-                    create_datasource.create()
+                    datasource.create()
                     mock_log.assert_called_with(
                         'error',
                         'monitoring_integration',
